requirements.txt
configparser==5.0.1

# missed tickets script requirements
amqp==5.0.5
anyjson==0.3.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.14
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5

# apibot requirements
redis==3.5.3

# snow to IRIS requirements
suds-jurko==0.6
suds-py3==1.4.4.1

# journal
DateTime==4.3
pika==1.2.0
pytz==2021.1
PyYAML==5.4.1
zope.interface==5.2.0

# IRIS connection
pyodbc==4.0.30